// apps/web/src/app/api/model/projectSummary.ts
/**
 * Talos API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectStatus } from './projectStatus';


export interface ProjectSummary {
    id: string;
    name: string;
    slug: string;
    repoUrl: string;
    defaultBranch: string;
    stackType: string;
    status: ProjectStatus;
    createdAt: string;
    updatedAt: string;
}
export namespace ProjectSummary {
}